Started migrating DatabaseManager to DataStore
