+++ elide-datastore/elide-datastore-hibernate3/src/test/java/com/yahoo/elide/datastores/hibernate3/HibernateStoreTest.java
@@ -1,0 +1,16 @@
+/*
+ * Copyright 2015, Yahoo Inc.
+ * Licensed under the Apache License, Version 2.0
+ * See LICENSE file in project root for terms.
+ */
+package com.yahoo.elide.datastores.hibernate3;
+
+import org.testng.annotations.Test;
+
+public class HibernateStoreTest {
+    // TODO: Create tests for our manager
+    @Test
+    public void todo() {
+        System.err.println("Still need to write tests for hibernate3 manager!");
+    }
+}
