+++ elide-datastore/elide-datastore-hibernate5/src/test/java/com/yahoo/elide/datastores/hibernate5/HibernateStoreTest.java
@@ -1,0 +1,11 @@
+package com.yahoo.elide.datastores.hibernate5;
+
+import org.testng.annotations.Test;
+
+public class HibernateStoreTest {
+    // TODO: Create tests for our manager
+    @Test
+    public void todo() {
+        System.err.println("Still need to write tests for hibernate3 manager!");
+    }
+}
