+++ elide-datastore/elide-datastore-inmemorydb/src/test/java/com/yahoo/elide/example/beans/NonEntity.java
