--- elide-dbmanager/elide-dbmanager-hibernate3/src/test/java/com/yahoo/elide/dbmanagers/hibernate3/HibernateManagerTest.java
@@ -1,18 +1,0 @@
-/*
- * Copyright 2015, Yahoo Inc.
- * Licensed under the Apache License, Version 2.0
- * See LICENSE file in project root for terms.
- */
-package com.yahoo.elide.dbmanagers.hibernate3;
-
-import org.testng.annotations.Test;
-
-import static org.testng.Assert.*;
-
-public class HibernateManagerTest {
-    // TODO: Create tests for our manager
-    @Test
-    public void todo() {
-        System.err.println("Still need to write tests for hibernate3 manager!");
-    }
-}
