--- elide-dbmanager/elide-dbmanager-hibernate5/src/test/java/com/yahoo/elide/dbmanagers/hibernate5/HibernateManagerTest.java
@@ -1,13 +1,0 @@
-package com.yahoo.elide.dbmanagers.hibernate5;
-
-import org.testng.annotations.Test;
-
-import static org.testng.Assert.*;
-
-public class HibernateManagerTest {
-    // TODO: Create tests for our manager
-    @Test
-    public void todo() {
-        System.err.println("Still need to write tests for hibernate3 manager!");
-    }
-}
