--- elide-dbmanager/pom.xml
@@ -1,60 +1,0 @@
-<!--
-  ~ Copyright 2015, Yahoo Inc.
-  ~ Licensed under the Apache License, Version 2.0
-  ~ See LICENSE file in project root for terms.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>elide-dbmanager-parent-pom</artifactId>
-    <packaging>pom</packaging>
-    <name>Elide Database Manager Parent Pom</name>
-    <description>Parent pom for database managers</description>
-    <url>https://github.com/yahoo/elide</url>
-    <parent>
-        <groupId>com.yahoo.elide</groupId>
-        <artifactId>elide-parent-pom</artifactId>
-        <version>1.0.0.6-SNAPSHOT</version>
-    </parent>
-
-    <licenses>
-        <license>
-            <name>The Apache Software License, Version 2.0</name>
-            <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
-            <distribution>repo</distribution>
-        </license>
-    </licenses>
-
-    <developers>
-        <developer>
-            <name>Yahoo Inc.</name>
-            <url>https://github.com/yahoo</url>
-        </developer>
-    </developers>
-
-    <scm>
-        <developerConnection>scm:git:ssh://git@github.com/yahoo/elide.git</developerConnection>
-        <url>https://github.com/yahoo/elide.git</url>
-        <tag>HEAD</tag>
-    </scm>
-
-    <properties>
-        <parent.pom.dir>${project.basedir}/../..</parent.pom.dir>
-    </properties>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>com.yahoo.elide</groupId>
-                <artifactId>elide-core</artifactId>
-                <version>1.0.0.6-SNAPSHOT</version>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-    <modules>
-        <module>elide-dbmanager-hibernate5</module>
-        <module>elide-dbmanager-hibernate3</module>
-        <module>elide-dbmanager-inmemorydb</module>
-    </modules>
-</project>
