+++ elide-example/elide-hibernate3-mysql-example/src/main/java/com/yahoo/elide/example/hibernate3/ElideResourceConfig.java
@@ -2,7 +2,7 @@ package com.yahoo.elide.example.hibernate3;
 
 import com.yahoo.elide.Elide;
 import com.yahoo.elide.audit.Slf4jLogger;
-import com.yahoo.elide.dbmanagers.hibernate3.HibernateManager;
+import com.yahoo.elide.datastores.hibernate3.HibernateStore;
 import com.yahoo.elide.resources.JsonApiEndpoint;
 import org.glassfish.hk2.utilities.binding.AbstractBinder;
 import org.glassfish.jersey.server.ResourceConfig;
@@ -21,7 +21,7 @@ public class ElideResourceConfig extends ResourceConfig {
                 bind(noUserFn).to(JsonApiEndpoint.DefaultOpaqueUserFunction.class).named("elideUserExtractionFunction");
 
                 SessionFactory sessionFactory = new Configuration().configure().buildSessionFactory();
-                Elide elide = new Elide(new Slf4jLogger(), new HibernateManager(sessionFactory));
+                Elide elide = new Elide(new Slf4jLogger(), new HibernateStore(sessionFactory));
                 bind(elide).to(Elide.class).named("elide");
             }
         });
