+++ elide-integration-tests/pom.xml
@@ -40,7 +40,7 @@
         </dependency>
         <dependency>
             <groupId>com.yahoo.elide</groupId>
-            <artifactId>elide-dbmanager-hibernate3</artifactId>
+            <artifactId>elide-datastore-hibernate3</artifactId>
             <version>1.0.0.6-SNAPSHOT</version>
         </dependency>
